requirements.txt
Django>=5.0,<5.1
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
whitenoise # -> upgrade to django-storages
django-allauth
django-allauth-ui
django-widget-tweaks
slippers
pandas
mistralai
pdf2image
